% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{eq_clean_data}
\usage{
eq_clean_data(input)
}
\arguments{
\item{input}{set Dataset depicting the raw NOAA data frame.}
}
\value{
This function returns a dataframe with the NOAA data frame cleaned.
}
\description{
this function takes raw NOAA data frame and returns a clean data
frame with a date column created by uniting the year, month, day
and converting it to the Date class,
LATITUDE and LONGITUDE columns converted to numeric class
}
\examples{
\dontrun{
input = read_delim("data/signif.txt", delim = "\\t")
dataset = eq_clean_data(input)
}
}
