% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_create_label}
\alias{eq_create_label}
\title{eq_create_label}
\usage{
eq_create_label(map_data)
}
\arguments{
\item{map_data}{A cleaned data frame with data obtained from NOAA website}
}
\value{
An HTML label that can be used as the annotation text in the leaflet map.
}
\description{
Created description of the earthquake next to the graph
}
\examples{
\dontrun{
readr::read_delim("signif.txt", delim = "\\t") \%>\%
  eq_clean_data() \%>\%
  eq_location_clean() \%>\%
  dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(date) >= 2000) \%>\%
  dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
  eq_map(annot_col = "popup_text")
}

}
