% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_clean.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{eq_location_clean}
\usage{
eq_location_clean(data, column)
}
\arguments{
\item{data}{Dataset clean NOAA data frame.}

\item{column}{The column LOCATION_NAME}
}
\value{
This function returns the raw NOAA
data frame with the LOCATION_NAME columns cleaned
}
\description{
limpie la variable LOCATION_NAME y conviértala al primer formato en mayúsculas
}
\examples{

\dontrun{

dataset = read_delim("data/signif.txt", delim = "\\t")
dataset = eq_location_clean(data= input, column="LOCATION_NAME")

}
}
