% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{eq_map}
\usage{
eq_map(data_map, annot_col = "date")
}
\arguments{
\item{data_map}{A cleaned data frame with data  from NOAA website}

\item{annot_col}{The name of the column from the data to be use for annotation}
}
\value{
A map of the earthquakes epicenters and providing some annotations

#' @examples
\dontrun{
readr::read_delim("signif.txt", delim = "\t") %>%
eq_clean_data() %>%
  dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(date) >= 2000) %>%
  eq_map(annot_col = "date")
}
}
\description{
Takes an argument data containing the filtered data frame
with earthquakes to visualize.
The function maps the epicenters (LATITUDE/LONGITUDE)
and annotates each point with in pop up window containing
annotation data stored in a column of the data frame.
}
