% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{GeomTimeline}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Mapping argument to the ggplot layer function.}

\item{data}{A Dataframe Object containing the data to plot.}

\item{stat}{Stat argument to the ggplot layer function.}

\item{position}{Position argument to the ggplot layer function.}

\item{na.rm}{na.rm argument to the ggplot layer function.}

\item{show.legend}{show.legend argument to the ggplot layer function.}

\item{inherit.aes}{inherit.aes argument to the ggplot layer function.}

\item{...}{Extra Params.}
}
\value{
This function adds the lineGrobs and pointGrob into the current graphics device.
}
\description{
This function displays the dates of Earthquakes
by presenting each Earthquake, on a straight line,
as a point of varying sizes depending on their Ritcher scale value (based on the \code{size} variable). The Earthquakes are color coded by the death toll it caused (\code{colour}). If the \code{y} variable is supplied, the Earthquakes will be displayed in multiple horizontal lines,
each line represnting the quakes for each Country.
}
\examples{

\dontrun{

dataset = readr::read_delim("data/signif.txt", delim = "\\t") \%>\%
  eq_clean_data() \%>\% eq_location_clean("LOCATION_NAME")\%>\%
  filter(COUNTRY \%in\% c("USA", "CHINA"), YEAR > 2000)

ggplot2::ggplot(input, ggplot2::aes(date,
                                      colour = TOTAL_DEATHS,
                                      y = COUNTRY,
                                      size = as.numeric(EQ_PRIMARY) )) +
  geom_timeline() +
  ggplot2::guides(size = ggplot2::guide_legend(title = "Ritcher scale value")) +
  ggplot2::scale_colour_continuous(name = "# of DEATHS") +
  ggplot2::theme_classic()

}

}
