% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{geom_timeline_label}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, n_max = 5, ...)
}
\arguments{
\item{mapping}{Mapping argument to the ggplot layer function.}

\item{data}{A Dataframe Object containing the data to plot.}

\item{stat}{Stat argument to the ggplot layer function.}

\item{position}{Position argument to the ggplot layer function.}

\item{na.rm}{na.rm argument to the ggplot layer function.}

\item{show.legend}{show.legend argument to the ggplot layer function.}

\item{inherit.aes}{inherit.aes argument to the ggplot layer function.}

\item{n_max}{maximum number of tags}

\item{...}{Extra Params.}
}
\value{
for adding annotations to the earthquake data
}
\description{
GeomTimeLineLabel Geom coding
}
\examples{

 \dontrun{

filename <- system.file("extdata", "signif.txt", package="earthquakeGeoms")
library(readr)
input <- readr::read_delim(filename, delim = "\\t")

 sample <- input \%>\%
 eq_clean_data() \%>\%
 eq_location_clean("LOCATION_NAME")\%>\%
 filter(YEAR >= 2000) \%>\%
 filter(COUNTRY \%in\% c("USA", "MEXICO"))

 ggplot(data = sample, aes(x=date,y=COUNTRY, label=LOCATION_NAME)) +
 geom_timeline() +
 geom_timeline_label(data=sample,aes(label=LOCATION_NAME),n_max=8) +
 scale_size_continuous(name = 'Richter scale value', guide = guide_legend(order = 1)) +
 geom_timeline(aes(y = COUNTRY, color=DEATHS, size = EQ_PRIMARY)) +
 scale_color_continuous(name = '# of Deaths', guide = guide_colorbar(order = 2)) +
 new_theme()
}

}
