% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{GeomTimeline}
\format{An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
GeomTimeline
}
\value{
foplotting a time line of earthquakes ranging from xmin to xmaxdates with a point for each earthquake
}
\description{
for plotting a time line of earthquakes ranging from xmin
to xmaxdates with a point for each earthquake.
Optional aesthetics include color, size, and alpha
(for transparency).
}
\details{
The data to be used with this geom must be downloaded and readed from NOAA site,
\url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}.
}
\section{Parameters to Geom}{


\code{data}: A Dataframe Object containing the data to plot. \cr \cr
\code{x}: Column of `data` depicting the x-coordinate of where each of the points will be located. \cr \cr
\code{size}: OPTIONAL. Column of `data` depicting size of each point in the pointsGrob displayed (if
missing, the first \code{nmax} labels will be displayed). \cr \cr
\code{y}: OPTIONAL. Column of `data` depicting the y-coordinate of where each of the points will be located. \cr \cr
\code{colour}: OPTIONAL. Column of `data` depicting the colour of each point in the pointGrob displayed. \cr \cr
\code{shape}: OPTIONAL. Column of 'data' depicting the shape of each of the points. \cr \cr
\code{alpha}: OPTIONAL. Column of `data` which will be used to apply the alpha to each point. \cr \cr
\code{fill}: OPTIONAL. Color 'string' (added outside of aes) to modify the default fill of the points (before applying \code{colour}). \cr \cr
\code{stroke}: OPTIONAL. Stroke of the image (need not be used). \cr \cr
}

\examples{

\dontrun{

filename <- system.file("extdata", "signif.txt", package="earthquakeGeoms")
library(readr)
input <- readr::read_delim(filename, delim = "\\t")

 sample <- input \%>\%
 eq_clean_data() \%>\%
 eq_location_clean("LOCATION_NAME")\%>\%
 filter(YEAR >= 2000) \%>\%
 filter(COUNTRY \%in\% c("USA", "MEXICO"))

ggplot2::ggplot(data=sample, aes(x = date, y = COUNTRY, color = DEATHS, size = EQ_PRIMARY)) +
 geom_timeline() +
 scale_size_continuous(name = 'Richter scale value', guide = guide_legend(order = 1)) +
 scale_color_continuous(name = '# of Deaths', guide = guide_colorbar(order = 2))

}

}
\keyword{datasets}
