% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\docType{data}
\name{GeomTimeLineLabel}
\alias{GeomTimeLineLabel}
\title{GeomTimeLineLabel}
\format{An object of class \code{GeomTimeLineLabel} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
GeomTimeLineLabel
}
\value{
for adding annotations to the earthquake data
}
\description{
adding annotations to the earthquake data.
This geom adds a vertical line to each data point with a text annotation (e.g.
the location of the earthquake) attached to each line.
}
\section{Parameters to Geom}{


 \code{x}: Time variable
 \code{label}: Factor for adding annotations to the earthquake data
 \code{y}: Factor indicating some stratification
 \code{n_max}: Number of earthquakes to subset according to their magnitude (EQ_PRIMARY)
}

\examples{

\dontrun{

filename <- system.file("extdata", "signif.txt", package="earthquakeGeoms")
library(readr)
input <- readr::read_delim(filename, delim = "\\t")

 sample <- input \%>\%
 eq_clean_data() \%>\%
 eq_location_clean("LOCATION_NAME")\%>\%
 filter(YEAR >= 2000) \%>\%
 filter(COUNTRY \%in\% c("USA", "MEXICO"))

 ggplot(data = sample, aes(x=date,y=COUNTRY, label=LOCATION_NAME)) +
 geom_timeline() +
 geom_timeline_label(data=sample,aes(label=LOCATION_NAME),n_max=8) +
 scale_size_continuous(name = 'Richter scale value', guide = guide_legend(order = 1)) +
 geom_timeline(aes(y = COUNTRY, color=DEATHS, size = EQ_PRIMARY)) +
 scale_color_continuous(name = '# of Deaths', guide = guide_colorbar(order = 2)) +
 new_theme()
}

}
\keyword{datasets}
