% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{new_theme}
\alias{new_theme}
\title{new_theme}
\usage{
new_theme(...)
}
\arguments{
\item{...}{Extra Params.}
}
\value{
This function adds the lineGrobs and pointGrob into the current graphics device.
}
\description{
Created a new theme tovisualizes ggplots
}
\examples{

\dontrun{

filename <- system.file("extdata", "signif.txt", package="earthquakeGeoms")
library(readr)
input <- readr::read_delim(filename, delim = "\\t")

 sample <- input \%>\%
 eq_clean_data() \%>\%
 eq_location_clean("LOCATION_NAME")\%>\%
 filter(YEAR >= 2000) \%>\%
 filter(COUNTRY \%in\% c("USA", "MEXICO"))

ggplot2::ggplot(data=sample, aes(x = date, y = COUNTRY, color = DEATHS, size = EQ_PRIMARY)) +
 geom_timeline() +
 scale_size_continuous(name = 'Richter scale value', guide = guide_legend(order = 1)) +
 scale_color_continuous(name = '# of Deaths', guide = guide_colorbar(order = 2))+
 new_theme()
}

}
